因爲以下三個理由，決定加上ArgumentMarshaler 這個 class，先加一個空殼
First, each argument type required some way to parse its schema element in order to select the HashMap for that type.
Next, each argument type needed to be parsed in the command-line strings and converted to its true type.
Finally, each argument type needed a getXXX method so that it could be returned to the caller as its true type.