把 setBoolean 的地方改成新增的 set 方法, 注意！那個string 參數, boolean 用不到, 但其他人用得到
And finally I replaced the call to setBoolean with a call to set.

Notice that the abstract set function takes a String argument, but the implementation in the BooleanArgumentMarshaller does not use it.
I put that argument in there because I knew that the StringArgumentMarshaller and IntegerArgumentMarshaller would use it.