加一個 ClassCastException 的 error handing
This last change might have been a surprise.
Why did I suddenly decide to deal with the ClassCastException?
The reason is that I have a set of unit tests and a separate set of acceptance tests written in FitNesse.
It turns out that the FitNesse tests made sure that if you called getBoolean on a nonboolean argument, you got a false.
The unit tests did not. Up to this point I had only been running the unit tests.